Refactor Supabase integration: update anon key, consolidate file upload logic, and enhance calendar event generation
